requirements.txt
asgiref==3.7.2
Babel==2.14.0
beautifulsoup4==4.12.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2023.10
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.2
django-crispy-forms==1.11.0
django-easy-audit==1.3.5
django-health-check==3.17.0
django-phonenumber-field==7.2.0
idna==3.4
oauthlib==3.2.2
phonenumbers==8.13.26
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.1.0
